#76 Rename module (search → arxiv)
