+++ arxiv_post/__init__.py
@@ -5,8 +5,8 @@ __version__ = "0.5.0"
 
 # submodules
 from . import apps
+from . import arxiv
 from . import article
 from . import cli
 from . import constants
 from . import deepl
-from . import search
